+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/manifest/DashManifestParser.java
@@ -47,6 +47,7 @@ import androidx.media3.exoplayer.upstream.ParsingLoadable;
 import androidx.media3.extractor.metadata.emsg.EventMessage;
 import androidx.media3.extractor.mp4.PsshAtomUtil;
 import com.google.common.base.Ascii;
+import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import java.io.ByteArrayOutputStream;
@@ -76,6 +77,20 @@ public class DashManifestParser extends DefaultHandler
   private static final Pattern CEA_608_ACCESSIBILITY_PATTERN = Pattern.compile("CC([1-4])=.*");
   private static final Pattern CEA_708_ACCESSIBILITY_PATTERN =
       Pattern.compile("([1-9]|[1-5][0-9]|6[0-3])=.*");
+
+  /**
+   * Maps the value attribute of an AudioChannelConfiguration with schemeIdUri
+   * "tag:dolby.com,2015:dash:audio_channel_configuration:2015" to the corresponding number of
+   * channels for each speaker group, as defined by ETSI TS 103 190-2 v1.2.1 clause G.3.
+   *
+   * <p>Table A.27 in ETSI TS 103 190-2 v1.2.1 defines the speaker counts for each speaker group
+   * index. The channel count is calculated by summing the speaker counts for the present indexes.
+   *
+   * <p>For example, a value of "0x3" (binary "11") indicates the presence of speaker groups 0 and
+   * 1. This maps to 2 + 1 = 3 channels.
+   */
+  private static final int[] DOLBY_AC4_CHANNEL_CONFIGURATION_MAPPING =
+      new int[] {2, 1, 2, 2, 2, 2, 1, 2, 2, 1, 1, 1, 1, 2, 1, 1, 2, 2, 2};
 
   /**
    * Maps the value attribute of an AudioChannelConfiguration with schemeIdUri
@@ -441,7 +456,7 @@ public class DashManifestParser extends DefaultHandler
       } else if (XmlPullParserUtil.isStartTag(xpp, "Role")) {
         roleDescriptors.add(parseDescriptor(xpp, "Role"));
       } else if (XmlPullParserUtil.isStartTag(xpp, "AudioChannelConfiguration")) {
-        audioChannels = parseAudioChannelConfiguration(xpp);
+        audioChannels = parseAudioChannelConfiguration(xpp, codecs);
       } else if (XmlPullParserUtil.isStartTag(xpp, "Accessibility")) {
         accessibilityDescriptors.add(parseDescriptor(xpp, "Accessibility"));
       } else if (XmlPullParserUtil.isStartTag(xpp, "EssentialProperty")) {
@@ -720,7 +735,7 @@ public class DashManifestParser extends DefaultHandler
         }
         baseUrls.addAll(parseBaseUrl(xpp, parentBaseUrls, dvbProfileDeclared));
       } else if (XmlPullParserUtil.isStartTag(xpp, "AudioChannelConfiguration")) {
-        audioChannels = parseAudioChannelConfiguration(xpp);
+        audioChannels = parseAudioChannelConfiguration(xpp, codecs);
       } else if (XmlPullParserUtil.isStartTag(xpp, "SegmentBase")) {
         segmentBase = parseSegmentBase(xpp, (SingleSegmentBase) segmentBase);
       } else if (XmlPullParserUtil.isStartTag(xpp, "SegmentList")) {
@@ -1487,7 +1502,7 @@ public class DashManifestParser extends DefaultHandler
 
   // AudioChannelConfiguration parsing.
 
-  protected int parseAudioChannelConfiguration(XmlPullParser xpp)
+  protected int parseAudioChannelConfiguration(XmlPullParser xpp, String codecs)
       throws XmlPullParserException, IOException {
     String schemeIdUri = parseString(xpp, "schemeIdUri", null);
     int audioChannels;
@@ -1508,6 +1523,9 @@ public class DashManifestParser extends DefaultHandler
       case "tag:dolby.com,2014:dash:audio_channel_configuration:2011":
       case "urn:dolby:dash:audio_channel_configuration:2011":
         audioChannels = parseDolbyChannelConfiguration(xpp);
+        break;
+      case "tag:dolby.com,2015:dash:audio_channel_configuration:2015":
+        audioChannels = parseDolbyAC4ChannelConfiguration(xpp, codecs);
         break;
       default:
         audioChannels = Format.NO_VALUE;
@@ -2042,6 +2060,65 @@ public class DashManifestParser extends DefaultHandler
       default:
         return Format.NO_VALUE;
     }
+  }
+
+  /**
+   * Parses the number of channels from the value attribute of an AudioChannelConfiguration with
+   * schemeIdUri "tag:dolby.com,2015:dash:audio_channel_configuration:2015" as defined by table A.27
+   * in ETSI TS 103 190-2 v1.2.1.
+   *
+   * @param xpp The parser from which to read.
+   * @param codecs The codecs string from the current element being parsed.
+   * @return The parsed number of channels, or {@link Format#NO_VALUE} if the channel count could
+   *     not be parsed.
+   */
+  protected static int parseDolbyAC4ChannelConfiguration(XmlPullParser xpp, String codecs) {
+    @Nullable String value = xpp.getAttributeValue(null, "value");
+    // The value attribute must be set to a six-digit uppercase hexadecimal string.
+    if (value == null || value.length() != 6) {
+      return Format.NO_VALUE;
+    }
+
+    // Mask for bit 23, indicating object-based audio (See E.10.14 presentation_channel_mask_v1).
+    int objectBasedAudioFlag = 1 << 23;
+    int ac4ChannelMask = Integer.parseInt(value, /* radix= */ 16);
+    if ((ac4ChannelMask & objectBasedAudioFlag) != 0) {
+      // object-based audio content.
+      return parseDolbyAc4ObjectBasedChannelConfiguration(codecs);
+    }
+    // channel-based audio content.
+    return parseDolbyAc4ChannelBasedChannelConfiguration(ac4ChannelMask);
+  }
+
+  private static int parseDolbyAc4ObjectBasedChannelConfiguration(String codecs) {
+    String[] codecList = Util.splitCodecs(codecs);
+    if (codecList.length == 0) {
+      return Format.NO_VALUE;
+    }
+
+    // The AC-4 codec string format is "ac-4.xx.yy.zz", where zz is presentation level.
+    List<String> parts = Splitter.on('.').splitToList(Ascii.toLowerCase(codecList[0].trim()));
+    if (parts.size() != 4 || !parts.get(0).equals("ac-4")) {
+      return Format.NO_VALUE;
+    }
+
+    switch (parts.get(3)) {
+      case "03":
+        return 18; // AC-4 Level 3 object-based content is mapped to 17.1 channels.
+      case "04":
+        return 21; // AC-4 Level 4 object-based content is mapped to 20.1 channels.
+      default:
+        return Format.NO_VALUE;
+    }
+  }
+
+  private static int parseDolbyAc4ChannelBasedChannelConfiguration(int ac4ChannelMask) {
+    // Bits 0...18 indicate the presence of individual channel groups.
+    int channelCount = 0;
+    for (int i = 0; i < DOLBY_AC4_CHANNEL_CONFIGURATION_MAPPING.length; i++) {
+      channelCount += ((ac4ChannelMask >> i) & 0x1) * DOLBY_AC4_CHANNEL_CONFIGURATION_MAPPING[i];
+    }
+    return channelCount == 0 ? Format.NO_VALUE : channelCount;
   }
 
   protected static long parseLastSegmentNumberSupplementalProperty(